requirements.txt
apache-airflow-providers-amazon==9.2.0
apache-airflow-providers-apache-spark==5.0.1
apache-airflow-providers-common-compat==1.3.0
boto3==1.36.3
pandas==2.1.4
pyspark==3.5.5
s3fs==2025.3.0